refactors the way filters and chains are built
