+++ lib/row_filter/filter.ex
@@ -2,8 +2,7 @@ defmodule Bigtable.RowFilter.Filter do
   @moduledoc """
   Behavior for creating and applying `Google.Bigtable.V2.RowFilter`
   """
-  alias Google.Bigtable.V2.{ReadRowsRequest, RowFilter}
+  alias Google.Bigtable.V2.RowFilter
 
-  @callback apply_filter(ReadRowsRequest.t(), any()) :: ReadRowsRequest.t()
   @callback build_filter(any()) :: RowFilter.t()
 end
